user now stays logged in, authentication setup for projects

+++ backend/routes/projects.js
@@ -1,26 +1,30 @@
 const router = require("express").Router()
+const auth = require("../middleware/auth")
 let Project = require("../models/Project.js")
 
-router.route("/").get((req, res) => {
-  Project.find()
-    .then(projects => res.json(projects))
-    .catch(err => res.status(400).json("Error: " + err))
+router.get("/", auth, async (req, res) => {
+  const projects = await Project.find({ userId: req.user })
+  res.json(projects)
 })
 
-router.route("/create").post((req, res) => {
-  const owner = req.body.owner
-  const title = req.body.title
-  const description = req.body.description
-  const steps = req.body.steps
-  const date = Date.parse(req.body.date)
+router.post("/create", auth, async (req, res) => {
+  try {
+    const { title, description, steps } = req.body
 
-  const newProject = new Project({
-    owner,
-    title,
-    description,
-    steps,
-    date
-  })
+    const date = Date.parse(req.body.date)
+
+    const newProject = new Project({
+      title,
+      description,
+      steps,
+      date,
+      userId: req.user
+    })
+    const savedProject = await newProject.save()
+    res.json(savedProject)
+  } catch (e) {
+    res.status(500).json({ error: e.message })
+  }
 
   newProject.save().then(() => res.json("Project added!").catch(err => res.status(400).json("Error: " + err)))
 })
@@ -31,16 +35,21 @@ router.route("/:id").get((req, res) => {
     .catch(err => res.status(400).json("Error: " + err))
 })
 
-router.route("/:id").delete((req, res) => {
-  Project.findByIdAndDelete(req.params.id)
-    .then(project => res.json(project))
-    .catch(err => res.status(400).json("Error: " + err))
+router.delete("/:id", auth, async (req, res) => {
+  const project = await Project.findOne({ userId: req.user, _id: req.params.id })
+  if (!project)
+    return res.status(400).json({
+      msg: "No todo found with this ID that belongs to the currently logged in user."
+    })
+
+  const deletedProject = await Project.findByIdAndDelete(req.params.id)
+  res.json(deletedProject)
 })
 
 router.route("/update/:id").post((req, res) => {
   Project.findById(req.params.id)
     .then(project => {
-      project.username = req.body.username
+      project.title = req.body.title
       project.description = req.body.description
       project.duration = Number(req.body.duration)
       project.date = Date.parse(req.body.date)
